Fixes: adding timeout delay, renaming player component
