+++ project/src/components/video-player/video-player.tsx
