+++ project/src/const.ts
@@ -15,5 +15,6 @@ export enum AuthStatus {
 
 export const SHOWN_FILM_LIMIT = 8;
 export const MAX_RATE = 10;
+export const PREVIEW_TIMEOUT = 1000;
 
 export const MOVIE_REF = 'https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm';
